refactor(recorder): extract TerminalCollector from ObservationRegistry
Move PendingExecution, MAX_OUTPUT_CHARS, the _pendingExecutions map,
shellExecStart/End listener registration, _collectExecutionOutput, and
_emitTerminalCommand into a new TerminalCollector class in
observation/terminalCollector.ts.

ObservationRegistry instantiates TerminalCollector via constructor injection
and delegates all three teardown paths (flushDebouncesForEnd,
discardDebouncesForConsentDowngrade, disposeSubscriptions) to
abortAllPending(). The enable() block is replaced with a single
_terminalCollector.register() call.

Behavior-preserving: the 5 characterization tests from the prior commit
pass without modification beyond the whitebox helper bodies, which now
reach through _terminalCollector instead of _observation directly.
